--- docs/exampleSetup.ts
@@ -1,40 +1,0 @@
-const dbConnection = new DataSource({
-  type: 'sqlite',
-  database: DATABASE_FILE,
-  synchronize: false,
-  migrations,
-  migrationsRun: true,
-  logging: ['error', 'info', 'warn'],
-  entities: Entities,
-}).initialize()
-
-export const agent = createAgent<
-  IDIDManager & IKeyManager & IDataStore & IDataStoreORM & IResolver & ICredentialPlugin
->({
-  plugins: [
-    new KeyManager({
-      store: new KeyStore(dbConnection),
-      kms: {
-        local: new KeyManagementSystem(new PrivateKeyStore(dbConnection, new SecretBox(KMS_SECRET_KEY))),
-      },
-    }),
-    new DIDManager({
-      store: new DIDStore(dbConnection),
-      defaultProvider: 'did:ethr:sepolia',
-      providers: {
-        'did:ethr:sepolia': new EthrDIDProvider({
-          defaultKms: 'local',
-          network: 'sepolia',
-          rpcUrl: 'https://sepolia.infura.io/v3/' + INFURA_PROJECT_ID,
-        }),
-      },
-    }),
-    new DIDResolverPlugin({
-      resolver: new Resolver({
-        ...ethrDidResolver({ infuraProjectId: INFURA_PROJECT_ID }),
-        ...webDidResolver(),
-      }),
-    }),
-    new CredentialPlugin(),
-  ],
-})
