+++ src/actors/issuerService.ts
@@ -1,6 +1,6 @@
 import { agentMediator} from '../veramoAgentMediator.js'
-import { agent as agentClient1} from '../veramoAgentClient1.js'
-import { agent as agentClient2} from '../veramoAgentClient2.js'
+import { agent as agentClient1} from '../veramoAgentHolder1.js'
+import { agent as agentClient2} from '../veramoAgentHolder2.js'
 
 export async function createVC(agent: any) {
   const identifier = await agent.didManagerGetByAlias({ alias: 'default' })
