+++ src/actors/verifierService.ts
@@ -1,6 +1,6 @@
 import { agentMediator} from '../veramoAgentMediator.js'
-import { agent as agentClient1} from '../veramoAgentClient1.js'
-import { agent as agentClient2} from '../veramoAgentClient2.js'
+import { agent as agentClient1} from '../veramoAgentHolder1.js'
+import { agent as agentClient2} from '../veramoAgentHolder2.js'
 
 export async function verifyVC(agent: any, credential: any) {
   const result = await agent.verifyCredential({
