--- test/testPostman.txt
@@ -1,12 +1,0 @@
-// Mediate request
-
-curl -X POST -H "Content-Type: application/json" \
--d '{
-  "type": "https://didcomm.org/coordinate-mediation/3.0/mediate-request",
-  "from": "did:ethr:sepolia:0x0226318a9fc4d308e32ac76663f92db890e73df6ca25206c05b4b5706413c8f983",
-  "to": "did:ethr:sepolia:0x0208c168457bcbed44c85c2805fff6e52d345a13cc51e33c447c741c08652607c9",
-  "id": "1234",
-  "created_time": "2025-09-29T18:43:09.693Z",
-  "body": {}
-}' \
-http://localhost:3000/didcomm
